requirements.txt
streamlit==1.44.1
openai==1.75.0
langchain==0.3.23
langchain-core==0.3.54
langchain-community
python-dotenv==1.1.0
